Work on classic theme handling multiple tutors
The package arrayjob has been imposed in the TeX local definition of
the user to handle multiple tutors. Its syntax is very intuitive. The
package xifthen was used in the classic theme to test optional argument
for exercise sheet title and to use the \whiledo command in the header
to print multiple tutors.