Add first implementation of -S option
Implementing the solution functionality, a lot of the code used to make
the exercises sheets could be reused and it was therefore extracted to
the AuxiliaryFunctions file. There, some functions that are almost
identical for exercises and solutions have been provided with an
argument which is either EXERCISE or SOLUTION in order to implement the
corresponding logic inside. Probably, there is a more elegant way to do
so (like extract the if in a function which call either a
function_SOLUTION or a function_EXERCISE), but for the moment it is
fine to do in this way. Still some code should be added on the latex
side (e.g. change title of sheet and title of solutions).