Treat theme as other latex files with blocks
So far, the theme was directly sourced in latex using the input
command, but it was also containing a PACKAGES block which was
extracted (so the packages were present twice in the final latex
source). Now it has a PACKAGES and a DEFINITIONS block, which are
extracted and put in the auxiliary files, as it is done for exercises
or for solutions.