Work on command line options parser
Now, the options of the script are classified as primary and secondary.
Primary options are mutually exclusive and each primary has a bunch of
secondary that can be used together with it. The helper is structured
in blocks, such that is clear which secondary options belong to which
primary one. Internally, code duplication has been reduced as much as
possible and this has led to a somehow not transparent code.
In particular,
  1) local variables have been used from within called functions,
     without passing them around (fine in the same file, if not
     abused)
  2) some associative arrays where used to map short and long options
     as well as options and help instructions.
We postponed to the future any refactoring.