Implement way to easily use user defined theme
The philosophy about the theme was changed and now a theme file must be
present in the invoking directory. The option -t can be given only in
the setup phase and it allows the user to either set his/her own theme
as in use or to get the default proposed theme as theme in use.