Fix minor bug in command line option parser
When no primary option was given and still some secondary ones where
specified, the check on the scendary options was failing due to bash
syntax errors. This case was simply not considered, but it should give
to the user an understandable error. The situation is fixed now. Not
that I changed the way to get the last element of an array from
${ARRAY[-1]} to ${ARRAY: -1} since the first gives an error on empty
arrays, while the second just returns an empty string.