Make presence sheet mode more versatile
Now it is possible to use the Exercise Handler in the presence sheet
mode even thought no final sheet is available (no exercise is required
if they are in any case going to be hidden in the table). This might be
the case for a tutor who would like only to keep trace of presences.